requirements.txt
pandas
pyarrow
tqdm
IPython
nltk
pyspellchecker
